% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdDistribution-package.R
\docType{package}
\name{birdDistribution-package}
\alias{birdDistribution}
\alias{birdDistribution-package}
\title{birdDistribution: Model Bird Distributions}
\description{
Build bird distribution and abundance models in R. The package currently uses INLA to build Bayesian models of bird distribution and abundance in Canada. This is a work in progress
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LandSciTech/birdDistribution}
  \item Report bugs at \url{https://github.com/LandSciTech/birdDistribution/issues}
}

}
\author{
\strong{Maintainer}: Sarah Endicott \email{sarah.endicott@canada.ca} (\href{https://orcid.org/0000-0001-9644-5343}{ORCID})

Authors:
\itemize{
  \item David Isles \email{david.isles@ec.gc.ca}
}

}
\keyword{internal}
