% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{cut.fn}
\alias{cut.fn}
\title{Rasterize a series of spatial predictions (needed for plotting)}
\usage{
\method{cut}{fn}(
  df = NA,
  target_raster = NA,
  column_name = NA,
  lower_bound = NA,
  upper_bound = NA
)
}
\arguments{
\item{df}{dataframe of predictions}

\item{target_raster}{raster with desired structure eg resolution, crs etc}

\item{column_name}{column in \code{df} to rasterize}

\item{lower_bound, upper_bound}{upper and lower bounds for labels}
}
\description{
Rasterize a series of spatial predictions (needed for plotting)
}
