% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{do_res_plot}
\alias{do_res_plot}
\title{Build INLA prediction map}
\usage{
do_res_plot(
  pred_rast,
  title,
  subtitle,
  subsubtitle = "",
  samp_grid,
  bcr_poly,
  col_pal_fn,
  species_label,
  levs_nm,
  file_nm
)
}
\arguments{
\item{pred_rast}{Raster of predictions}

\item{title}{map title describing the metric shown}

\item{subtitle}{subtitle under title with more details}

\item{subsubtitle}{subtitle under subtitle eg with units or further description}

\item{samp_grid}{Points in sampling grid where species was or wasn't detected}

\item{col_pal_fn}{function to create colour palette}

\item{species_label}{species name which will be added to the map}

\item{levs_nm}{name of levels of colour ramp}

\item{file_nm}{file name where map should be saved}
}
\value{
saves the map to \code{file_nm}
}
\description{
Build INLA prediction map
}
