% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_extract_mod.R
\name{ee_extract}
\alias{ee_extract}
\title{Modified version of ee_extract function}
\usage{
ee_extract(
  x,
  y,
  fun = ee$Reducer$mean(),
  scale = NULL,
  sf = FALSE,
  via = "getInfo",
  container = "rgee_backup",
  lazy = FALSE,
  quiet = FALSE,
  ...
)
}
\description{
See Issue: https://github.com/r-spatial/rgee/issues/367
}
\details{
See \code{\link[rgee:ee_extract]{rgee::ee_extract()}} for documentation
}
