% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{evaluate_preds}
\alias{evaluate_preds}
\title{Evaluate model performance based on predicted vs observed count}
\usage{
evaluate_preds(pred, mod, sp_code, analysis_data)
}
\arguments{
\item{pred}{prediction data frame}

\item{mod}{fitted model}

\item{sp_code}{species code}

\item{analysis_data}{list containing analysis data, including \code{full_count_matrix}}
}
\value{
data frame of performance metrics
}
\description{
Compare predicted median abundance to observed count. Also summarise CRPS and log score if available
}
\examples{

test_dat <- readRDS(system.file("extdata", "analysis_data_test.rds", package = "birdDistribution"))

test_dat$all_surveys <- test_dat$all_surveys \%>\% mutate(Obs_Index = 1:n())

# takes awhile to run
if(FALSE){
# train on old data
  mod <- fit_inla(
    sp_code = test_dat$species_to_model$Species_Code_BSC[1],
    analysis_data = test_dat,
    proj_use = AEA_proj,
    study_poly = test_area,
    covariates = cov_df,
    mod_dir = tempdir(),
    save_mod = FALSE,
    train_dat_filter = "Date_Time < lubridate::ymd('2010-05-01')"
  )
# Predict on new data
  pred <- predict_inla(
    dat = test_dat$all_surveys \%>\% filter(Date_Time >= lubridate::ymd('2010-05-01')),
    analysis_data = test_dat,
    mod = mod,
    sp_code = test_dat$species_to_model$Species_Code_BSC[1],
    covariates = cov_df,
    do_crps = TRUE
  )

evaluate_preds(pred \%>\% mutate(Crossval_Fold = 1), mod,
               sp_code = test_dat$species_to_model$Species_Code_BSC[1],
               test_dat)
}

}
