% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{fit_inla}
\alias{fit_inla}
\title{Fit INLA model}
\usage{
fit_inla(
  sp_code,
  analysis_data,
  proj_use,
  study_poly,
  covariates,
  mod_dir = "data/derived-data/INLA_results/models/",
  train_dat_filter = "TRUE",
  save_mod = TRUE,
  file_name_bit = "all",
  bru_verbose = 4
)
}
\arguments{
\item{sp_code}{species four letter code}

\item{analysis_data}{list containing analysis data, including \code{full_count_matrix},
\code{all_surveys}, \code{species_ranges}, and \code{species_to_model}}

\item{proj_use}{projection/coordinate reference system to use. Note this
should have units of kms to avoid an extremely dense mesh}

\item{study_poly}{sf polygon of study area}

\item{covariates}{a data frame with columns covariate, model, mean, prec, beta
which is used to define the model formula.}

\item{mod_dir}{directory where the model object will be saved or loaded from
if it already exists}

\item{train_dat_filter}{a string that will be used to filter the input data,
the default will not filter anything}

\item{save_mod}{logical. Should the model object be saved?}

\item{file_name_bit}{suffix attached to the file name eg to identify
cross-validation fold}

\item{bru_verbose}{level of verbosity from bru. Lower number leads to less
output. See \code{\link[inlabru:bru_options]{inlabru::bru_options()}}}
}
\value{
An INLA object with the fit model
}
\description{
Fit a model of bird abundance using INLA
}
\examples{
# get test data set
test_dat <- readRDS(system.file("extdata", "analysis_data_test.rds", package = "birdDistribution"))

# Using km for projection units because it helps with INLA model calculations to have smaller numbers
AEA_proj <- "+proj=aea +lat_1=50 +lat_2=70 +lat_0=40 +lon_0=-106 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=km +no_defs "

# takes awhile to run
if(FALSE){
mod <- fit_inla(
  sp_code = test_dat$species_to_model$Species_Code_BSC[1],
  analysis_data = test_dat,
  proj_use = AEA_proj,
  study_poly = test_area,
  covariates = cov_df,
  mod_dir = tempdir(),
  save_mod = FALSE
)
}
}
