% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{get_dist_to_range}
\alias{get_dist_to_range}
\title{Get distance from point to ebird range}
\usage{
get_dist_to_range(sp_dat, sp_code, species_ranges)
}
\arguments{
\item{sp_dat}{combined species count and survey data}

\item{sp_code}{species four letter code}

\item{species_ranges}{A list containing ebird range data named by species code}
}
\value{
data frame with distance_from_range column added
}
\description{
Distance will be zero inside the species' range.
}
\examples{
library(dplyr)
library(sf)
test_dat <- readRDS(system.file("extdata", "analysis_data_test.rds", package = "birdDistribution"))
surv_pt <- test_dat$all_surveys \%>\%
  slice(1) \%>\%
  select(Survey_Type, Survey_Duration_Minutes)
get_dist_to_range(surv_pt, "WTSP", test_dat$species_ranges)
}
