% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{make_mesh}
\alias{make_mesh}
\title{Create a spatial mesh, which is used to fit the residual spatial field}
\usage{
make_mesh(poly, proj_use, max.edge = c(70000, 1e+05), cutoff = 30000)
}
\arguments{
\item{poly}{sf polygon of area to make mesh in. Eg the study area}

\item{proj_use}{projection/coordinate reference system to use. Note this
should have units of kms to avoid an extremely dense mesh}

\item{max.edge}{The largest allowed triangle edge length. See \code{\link[fmesher:fm_mesh_2d]{fmesher::fm_mesh_2d_inla()}} for details}

\item{cutoff}{The minimum allowed distance between points. See \code{\link[fmesher:fm_mesh_2d]{fmesher::fm_mesh_2d_inla()}} for details}
}
\description{
Create a spatial mesh, which is used to fit the residual spatial field
}
