% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{predict_inla}
\alias{predict_inla}
\title{Use a fit INLA model to generate predictions}
\usage{
predict_inla(dat, analysis_data, mod, sp_code, covariates, do_crps = TRUE)
}
\arguments{
\item{dat}{data to make predictions from}

\item{analysis_data}{list containing analysis data, including \code{full_count_matrix},
\code{all_surveys}, \code{species_ranges}, and \code{species_to_model}}

\item{mod}{fitted INLA model}

\item{sp_code}{species four letter code}

\item{covariates}{a data frame with columns covariate, model, mean, prec, beta
which is used to define the model formula.}

\item{do_crps}{logical. Should the Continuous Ranked Probability Score and
the Log score be calculated from the raw predictions?}
}
\value{
\code{dat} with columns added for predictions.
}
\description{
Generate predictions from a predictor data set and a fitted INLA model.
Predictions are summarised using the median, credible interval, and
coefficient of variation. The Continuous Ranked Probability Score and the Log
score are optionally calculated.
}
