% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{wt_dl_loc_year}
\alias{wt_dl_loc_year}
\title{Download location and year information from WildTrax}
\usage{
wt_dl_loc_year(proj_id, sens_id)
}
\arguments{
\item{proj_id}{project_id in wildTrax}

\item{sens_id}{type of sensor, either "ARU" or "point_count"}
}
\description{
Download location and year information from WildTrax
}
