% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_extract_mod.R
\name{ee_extract}
\alias{ee_extract}
\title{Modified version of ee_extract function}
\usage{
ee_extract(
  x,
  y,
  fun = rgee::ee$Reducer$mean(),
  scale = NULL,
  sf = FALSE,
  via = "getInfo",
  container = "rgee_backup",
  lazy = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{x}{ee$Image.}

\item{y}{ee$Geometry$*, ee$Feature, ee$FeatureCollection, sfc or sf objects.}

\item{fun}{ee$Reducer object. Function to summarize the values. The function
must take a single numeric value as an argument and return a single value.
See details.}

\item{scale}{A nominal scale given in meters of the Image projection to work in.
By default 1000.}

\item{sf}{Logical. Should the function return an sf object?}

\item{via}{Character. Method for exporting the image.
Three methods are available: "getInfo", "drive", "gcs".}

\item{container}{Character. Name of the folder ('drive') or bucket ('gcs')
to export the image into (ignore if \code{via} is not defined as "drive" or
"gcs").}

\item{lazy}{Logical. If TRUE, a \code{\link[future:sequential]{
future::sequential}} object is created to evaluate the task in the future.
Ignore if \code{via} is set as "getInfo". See details.}

\item{quiet}{Logical. Suppress info message.}

\item{...}{ee$Image$reduceRegions additional parameters. See
\code{ee_help(ee$Image$reduceRegions)} for more details.}
}
\description{
See Issue: https://github.com/r-spatial/rgee/issues/367
}
\details{
See \code{\link[rgee:ee_extract]{rgee::ee_extract()}} for documentation
}
