% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{get_QPAD_offsets}
\alias{get_QPAD_offsets}
\title{Get QPAD Offsets}
\usage{
get_QPAD_offsets(sp_dat, sp_code, offset_table)
}
\arguments{
\item{sp_dat}{combined species count and survey data}

\item{sp_code}{species four letter code}

\item{offset_table}{data frame of offset data}
}
\description{
Use a pre-generated table of EDR and cue rate from QPAD to calculate offsets
}
