% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{map_inla_preds}
\alias{map_inla_preds}
\title{Make maps of INLA model predictions}
\usage{
map_inla_preds(
  sp_code,
  analysis_data,
  preds,
  proj_use,
  atlas_squares,
  bcr_poly,
  study_poly,
  map_dir = "data/derived-data/INLA_results/maps/",
  train_dat_filter = "TRUE",
  file_name_bit = "all"
)
}
\arguments{
\item{sp_code}{species four letter code}

\item{analysis_data}{list containing analysis data, including \code{full_count_matrix},
\code{all_surveys}, \code{species_ranges}, and \code{species_to_model}}

\item{preds}{data frame of predictions}

\item{proj_use}{projection/coordinate reference system to use.}

\item{atlas_squares}{grid of squares to show predictions and observations in.}

\item{bcr_poly}{polygon of BCR boundaries to use in map.}

\item{study_poly}{sf polygon of study area.}

\item{map_dir}{directory where the map images should be saved}

\item{train_dat_filter}{a string that will be used to filter the input data,
the default will not filter anything}

\item{file_name_bit}{suffix attached to the file name eg to identify
cross-validation fold}
}
\description{
Makes several maps from different prediction outputs
}
\examples{
test_dat <- readRDS(system.file("extdata", "analysis_data_test.rds", package = "birdDistribution"))

# takes awhile to run
if(FALSE){
  mod <- fit_inla(
    sp_code = test_dat$species_to_model$Species_Code_BSC[1],
    analysis_data = test_dat,
    proj_use = AEA_proj,
    study_poly = test_area,
    covariates = cov_df,
    mod_dir = tempdir(),
    save_mod = FALSE
  )

  pred <- predict_inla(
    dat = test_dat$ONGrid,
    analysis_data = test_dat,
    mod = mod,
    sp_code = test_dat$species_to_model$Species_Code_BSC[1],
    covariates = cov_df,
    do_crps = FALSE
  )

dir_use <- tempdir()
map_inla_preds(
  sp_code = test_dat$species_to_model$Species_Code_BSC[1],
  analysis_data = test_dat,
  pred,
  proj_use = AEA_proj,
  study_poly = test_area,
  atlas_squares = atl_sq \%>\% sf::st_transform(AEA_proj),
  bcr_poly = bcr_poly,
  map_dir = dir_use
)

out_maps <- list.files(dir_use, pattern = "png$", full.names = TRUE)
}
}
