% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{predict_inla}
\alias{predict_inla}
\title{Use a fit INLA model to generate predictions}
\usage{
predict_inla(dat, analysis_data, mod, sp_code, covariates, do_crps = TRUE)
}
\arguments{
\item{dat}{data to make predictions from}

\item{analysis_data}{list containing analysis data, including \code{full_count_matrix},
\code{all_surveys}, \code{species_ranges}, and \code{species_to_model}}

\item{mod}{fitted INLA model}

\item{sp_code}{species four letter code}

\item{covariates}{a data frame with columns covariate, model, mean, prec, beta
which is used to define the model formula.}

\item{do_crps}{logical. Should the Continuous Ranked Probability Score and
the Log score be calculated from the raw predictions?}
}
\value{
\code{dat} with columns added for predictions.
}
\description{
Generate predictions from a predictor data set and a fitted INLA model.
Predictions are summarised using the median, credible interval, and
coefficient of variation. The Continuous Ranked Probability Score and the Log
score are optionally calculated.
}
\examples{

# get test data set
test_dat <- readRDS(system.file("extdata", "analysis_data_test.rds", package = "birdDistribution"))

# Using km for projection units because it helps with INLA model calculations to have smaller numbers
AEA_proj <- "+proj=aea +lat_1=50 +lat_2=70 +lat_0=40 +lon_0=-106 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=km +no_defs "

# takes awhile to run
if(FALSE){
  mod <- fit_inla(
    sp_code = test_dat$species_to_model$Species_Code_BSC[1],
    analysis_data = test_dat,
    proj_use = AEA_proj,
    study_poly = test_area,
    covariates = cov_df,
    mod_dir = tempdir(),
    save_mod = FALSE
  )

  pred <- predict_inla(
    dat = test_dat$ONGrid,
    analysis_data = test_dat,
    mod = mod,
    sp_code = test_dat$species_to_model$Species_Code_BSC[1],
    covariates = cov_df,
    do_crps = FALSE
  )
}
}
