% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_functions.R
\name{prep_sp_dat}
\alias{prep_sp_dat}
\title{Combine survey and count data for a species, optionally filter the data used}
\usage{
prep_sp_dat(
  analysis_data,
  sp_code,
  proj_use,
  train_dat_filter = "TRUE",
  survey_types = c("Point_Count", "ARU_SPT", "ARU_SPM")
)
}
\arguments{
\item{analysis_data}{list containing analysis data, including \code{full_count_matrix} and \code{all_surveys}}

\item{sp_code}{species four letter code}

\item{proj_use}{projection/coordinate reference system to use}

\item{train_dat_filter}{a string that will be used to filter the input data, the default will not filter anything}

\item{survey_types}{character vector of values in the \code{Survey_Type} column to keep}
}
\description{
Combine survey and count data for a species, optionally filter the data used
}
